+++ lib/advanced_bmi.dart
@@ -1,3 +1,5 @@
 library advanced_bmi;
 
-export 'src/main.dart';
+export 'src/bmi_controller.dart';
+export 'src/bmi_model.dart';
+export 'src/constants.dart';